move func VerifyPixels() from server package to verification package
