+++ server/broadcast.go
@@ -3,11 +3,9 @@ package server
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"math"
 	"math/big"
 	"net/url"
-	"os"
 	"strings"
 	"sync"
 
@@ -24,7 +22,6 @@ import (
 	"github.com/livepeer/go-livepeer/pm"
 	"github.com/livepeer/go-livepeer/verification"
 
-	"github.com/livepeer/lpms/ffmpeg"
 	"github.com/livepeer/lpms/stream"
 )
 
@@ -428,7 +425,7 @@ func transcodeSegment(cxn *rtmpConnection, seg *stream.HLSSegment, name string,
 		// Only run if a verifier is not being used
 		if sess.Sender != nil && verifier == nil {
 			go func() {
-				if err := verifyPixels(url, sess.BroadcasterOS, pixels); err != nil {
+				if err := verification.VerifyPixels(url, sess.BroadcasterOS, pixels); err != nil {
 					glog.Error(err)
 					cxn.sessManager.removeSession(sess)
 				}
@@ -600,53 +597,6 @@ func verify(verifier *verification.SegmentVerifier, cxn *rtmpConnection,
 		return nil
 	}
 	return err // possibly nil
-}
-
-func verifyPixels(fname string, bos drivers.OSSession, reportedPixels int64) error {
-	uri, err := url.ParseRequestURI(fname)
-	memOS, ok := bos.(*drivers.MemorySession)
-	// If the filename is a relative URI and the broadcaster is using local memory storage
-	// fetch the data and write it to a temp file
-	if err == nil && !uri.IsAbs() && ok {
-		tempfile, err := ioutil.TempFile("", common.RandName())
-		if err != nil {
-			return fmt.Errorf("error creating temp file for pixels verification: %v", err)
-		}
-		defer os.Remove(tempfile.Name())
-		defer tempfile.Close()
-
-		data := memOS.GetData(fname)
-		if data == nil {
-			return errors.New("error fetching data from local memory storage")
-		}
-
-		if _, err := tempfile.Write(memOS.GetData(fname)); err != nil {
-			return fmt.Errorf("error writing temp file for pixels verification: %v", err)
-		}
-
-		fname = tempfile.Name()
-	}
-
-	p, err := pixels(fname)
-	if err != nil {
-		return err
-	}
-
-	if p != reportedPixels {
-		return errors.New("mismatch between calculated and reported pixels")
-	}
-
-	return nil
-}
-
-func pixels(fname string) (int64, error) {
-	in := &ffmpeg.TranscodeOptionsIn{Fname: fname}
-	res, err := ffmpeg.Transcode3(in, nil)
-	if err != nil {
-		return 0, err
-	}
-
-	return res.Decoded.Pixels, nil
 }
 
 // Return an updated copy of the given session using the received transcode result
