+++ verification/verify.go
@@ -2,9 +2,15 @@ package verification
 
 import (
 	"errors"
+	"fmt"
+	"io/ioutil"
+	"net/url"
+	"os"
 	"sort"
 
+	"github.com/livepeer/go-livepeer/common"
 	"github.com/livepeer/go-livepeer/core"
+	"github.com/livepeer/go-livepeer/drivers"
 	"github.com/livepeer/go-livepeer/net"
 
 	"github.com/livepeer/lpms/ffmpeg"
@@ -101,6 +107,9 @@ func (sv *SegmentVerifier) Verify(params *Params) (*Params, error) {
 	// TODO sig checking; extract from broadcast.go
 
 	if sv.policy.Verifier == nil {
+		// check if the node is on-chain mode
+		if params != nil && params.Orchestrator != nil && params.Orchestrator.TicketParams != nil {
+		}
 		return nil, nil
 	}
 	// TODO Use policy sampling rate to determine whether to invoke verifier.
@@ -112,6 +121,10 @@ func (sv *SegmentVerifier) Verify(params *Params) (*Params, error) {
 		if len(res.Pixels) != len(params.Results.Segments) {
 			// TODO make allowances for the verification algos not doing
 			//      pixel counts themselves; adapt broadcast.go verifyPixels
+
+			// check if the node is on-chain mode
+			if params != nil && params.Orchestrator != nil && params.Orchestrator.TicketParams != nil {
+			}
 			err = ErrPixelsAbsent
 		}
 		for i := 0; err == nil && i < len(params.Results.Segments) && i < len(res.Pixels); i++ {
@@ -152,4 +165,51 @@ func (sv *SegmentVerifier) Verify(params *Params) (*Params, error) {
 func IsRetryable(err error) bool {
 	_, retryable := err.(Retryable)
 	return retryable
+}
+
+func VerifyPixels(fname string, bos drivers.OSSession, reportedPixels int64) error {
+	uri, err := url.ParseRequestURI(fname)
+	memOS, ok := bos.(*drivers.MemorySession)
+	// If the filename is a relative URI and the broadcaster is using local memory storage
+	// fetch the data and write it to a temp file
+	if err == nil && !uri.IsAbs() && ok {
+		tempfile, err := ioutil.TempFile("", common.RandName())
+		if err != nil {
+			return fmt.Errorf("error creating temp file for pixels verification: %v", err)
+		}
+		defer os.Remove(tempfile.Name())
+		defer tempfile.Close()
+
+		data := memOS.GetData(fname)
+		if data == nil {
+			return errors.New("error fetching data from local memory storage")
+		}
+
+		if _, err := tempfile.Write(memOS.GetData(fname)); err != nil {
+			return fmt.Errorf("error writing temp file for pixels verification: %v", err)
+		}
+
+		fname = tempfile.Name()
+	}
+
+	p, err := Pixels(fname)
+	if err != nil {
+		return err
+	}
+
+	if p != reportedPixels {
+		return errors.New("mismatch between calculated and reported pixels")
+	}
+
+	return nil
+}
+
+func Pixels(fname string) (int64, error) {
+	in := &ffmpeg.TranscodeOptionsIn{Fname: fname}
+	res, err := ffmpeg.Transcode3(in, nil)
+	if err != nil {
+		return 0, err
+	}
+
+	return res.Decoded.Pixels, nil
 }
